+++ validity/CHANGELOG.md
@@ -1,2 +1,3 @@
 Unreleased changes:
+* decorateList in Data.Validity
 * -0 is now a valid value for Double and Float.